Changed Start Configuration
Changed Start Configuration to more accurate represent when the server is ready for client connections.

old was "DungeonDB Start" but clients are not actually able to connect until the server says "Game server connected"

When "DungeonDB Start" appears, the server may still be generating or loading world files, and clients will not be able to connect if it is doing so. The world is done loading/generating when the message "Game server connected" appears.